// qwipo-mobile/src/types.ts
export type NotificationType = 'order' | 'promo' | 'stock' | 'system';

export interface AppNotification {
  id: string;
  type: NotificationType;
  icon: string;
  iconColor: string;
  title: string;
  body: string;
  time: string;
  unread: boolean;
}